requirements.txt
openai==1.41.1
orjson==3.10.7
pydantic==2.8.2
python-dotenv==1.0.1
tiktoken==0.7.0
webvtt-py==0.5.1


